+++ arviz/plots/backends/bokeh/posteriorplot.py
@@ -13,6 +13,7 @@ from ...plot_utils import (
     calculate_point_estimate,
     format_sig_figs,
     round_num,
+    vectorized_to_hex,
 )
 from .. import show_layout
 from . import backend_kwarg_defaults, create_axes_grid
@@ -37,6 +38,8 @@ def plot_posterior(
     textsize,
     ref_val,
     rope,
+    ref_val_color,
+    rope_color,
     labeller,
     kwargs,
     backend_kwargs,
@@ -87,6 +90,8 @@ def plot_posterior(
             linewidth=linewidth,
             ref_val=ref_val,
             rope=rope,
+            ref_val_color=ref_val_color,
+            rope_color=rope_color,
             ax_labelsize=ax_labelsize,
             **kwargs,
         )
@@ -117,6 +122,8 @@ def _plot_posterior_op(
     skipna,
     ref_val,
     rope,
+    ref_val_color,
+    rope_color,
     ax_labelsize,
     round_to: Optional[int] = None,
     **kwargs,
@@ -155,9 +162,20 @@ def _plot_posterior_op(
             val,
             format_as_percent(greater_than_ref_probability, 1),
         )
-        ax.line([val, val], [0, 0.8 * max_data], line_color="blue", line_alpha=0.65)
+        ax.line(
+            [val, val],
+            [0, 0.8 * max_data],
+            line_color=vectorized_to_hex(ref_val_color),
+            line_alpha=0.65,
+        )
 
-        ax.text(x=[values.mean()], y=[max_data * 0.6], text=[ref_in_posterior], text_align="center")
+        ax.text(
+            x=[values.mean()],
+            y=[max_data * 0.6],
+            text=[ref_in_posterior],
+            text_color=vectorized_to_hex(ref_val_color),
+            text_align="center",
+        )
 
     def display_rope(max_data):
         if rope is None:
@@ -185,15 +203,28 @@ def _plot_posterior_op(
             vals,
             (max_data * 0.02, max_data * 0.02),
             line_width=linewidth * 5,
-            line_color="red",
+            line_color=vectorized_to_hex(rope_color),
             line_alpha=0.7,
         )
-
+        probability_within_rope = ((values > vals[0]) & (values <= vals[1])).mean()
         text_props = dict(
-            text_font_size="{}pt".format(ax_labelsize), text_color="black", text_align="center"
+            text_color=vectorized_to_hex(rope_color),
+            text_align="center",
+        )
+        ax.text(
+            x=values.mean(),
+            y=[max_data * 0.45],
+            text=[f"{format_as_percent(probability_within_rope, 1)} in ROPE"],
+            **text_props,
         )
 
-        ax.text(x=vals, y=[max_data * 0.2, max_data * 0.2], text=rope_text, **text_props)
+        ax.text(
+            x=vals,
+            y=[max_data * 0.2, max_data * 0.2],
+            text_font_size="{}pt".format(ax_labelsize),
+            text=rope_text,
+            **text_props,
+        )
 
     def display_point_estimate(max_data):
         if not point_estimate: